+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [5.11.0](https://github.com/plivo/plivo-java/tree/v5.11.0) (2022-05-05)
+**Feature - List all recordings**
+- `fromNumber` and `toNumber` added to filtering param [List all recordings](https://www.plivo.com/docs/voice/api/recording#list-all-recordings)
+- `recordMinMemberCount` param added in [Add a participant to a multiparty call using API](https://www.plivo.com/docs/voice/api/multiparty-call/participants#add-a-participant)
+
 ## [5.10.0](https://github.com/plivo/plivo-java/tree/v5.10.0) (2022-03-25)
 **Feature - DialElement**
 - `confirmTimeout` parameter added to [The Dial element](https://www.plivo.com/docs/voice/xml/dial/)